requirements.txt
# CERBERUS-FANGS LANCELOTT - Unified Requirements
# Updated for Crush + CliWrap Integration and LangChain Support

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
jinja2==3.1.2
websockets==12.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Testing and Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
types-requests==2.31.0.10

# Core Dependencies
requests==2.31.0

# System & Process Management
psutil==5.9.8
pexpect==4.9.0
paramiko==3.4.0
fabric==3.2.2
colorama==0.4.6
click==8.1.3
rich==13.7.0
urllib3==1.26.13

# Enhanced Terminal and UI (for Crush orchestrator)
rich-click==1.7.1
textual==0.44.1
prompt-toolkit==3.0.41

# AI and Machine Learning Integration
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
openai>=1.6.0
anthropic>=0.8.0
transformers>=4.36.0
torch>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Firebase Integration
firebase-admin>=6.2.0
firebase-functions>=0.1.0
pyrebase4>=4.7.1
flask>=2.3.0
flask-cors>=4.0.0
google-cloud-firestore>=2.13.1
google-cloud-storage>=2.10.0
google-auth>=2.23.4
google-api-python-client>=2.108.0
sentry-sdk>=1.38.0
asyncio-mqtt>=0.11.0

# Configuration Management (YAML support for enhanced configs)
PyYAML>=6.0
ruamel.yaml>=0.18.0
cerberus>=1.3.4

# Enhanced Logging and Monitoring
loguru>=0.7.0
structlog>=23.2.0
prometheus-client>=0.19.0
grafana-client>=3.5.0

# Web Scraping and Parsing
beautifulsoup4>=4.11.2
lxml>=4.9.2
bs4
selenium>=4.15.0
playwright>=1.40.0

# Security and Cryptography
cryptography>=41.0.0
pyOpenSSL>=23.3.0
certifi>=2023.11.17
keyring>=24.3.0
secrets>=1.0.0

# Network and DNS
dnspython>=2.4.2
python-nmap>=0.7.1
netaddr>=0.9.0
aiohttp>=3.9.0
whois>=0.9.27
python-whois>=0.7.3
scapy>=2.5.0

# Database Support (Enhanced)
sqlite3-utils>=0.4.0
pymongo>=4.6.0
redis-py-cluster>=2.1.3

# File Management and Compression (for Crush integration)
patool>=1.12.1
rarfile>=4.1
py7zr>=0.20.8
zipfile38>=0.0.3

# Process and Command Management (for CliWrap integration)
subprocess-run>=0.6.0
shlex-quote>=0.0.3
ptyprocess>=0.7.0

# ARGUS Dependencies
aioquic>=0.9.20
gevent>=23.9.1
pyfiglet>=1.0.2
scapy>=2.5.0

# Kraken Dependencies
paramiko>=3.4.0
pywifi>=1.1.12
pyrad>=2.4
ldap3>=2.9.1

# RedTeam-ToolKit Dependencies
django>=4.2.7
gunicorn==20.1.0
bandit==1.7.5
bcrypt>=4.1.2
chardet==5.2.0
charset-normalizer>=3.3.2
contourpy>=1.2.0
cycler>=0.12.1
defusedxml>=0.7.1
dirsearch>=0.4.3
fonttools>=4.47.0
fpdf>=1.7.2
gitdb>=4.0.11
GitPython>=3.1.40
google>=3.0.0
greenlet>=3.0.2
humanize>=4.9.0
idna>=3.6
MarkupSafe>=2.1.3
matplotlib>=3.8.2
mypy-extensions>=1.0.0
networkx>=3.2.1
ntlm-auth>=1.5.0
packaging>=23.2
pbr>=6.0.0
Pillow>=10.1.0
proxy-db>=0.3.1
pycparser>=2.21
PyNaCl>=1.5.0
pyparsing>=3.1.1
PySocks>=1.7.1
python-dateutil>=2.8.2
python-decouple>=3.8
pytz>=2023.3
requests-mock>=1.11.0
requests-ntlm>=1.2.0
six>=1.16.0
smmap>=5.0.1
soupsieve>=2.5
sqlparse>=0.4.4
stevedore>=5.1.0
toml>=0.10.2
tomli>=2.0.1
typing_extensions>=4.8.0
uro>=0.0.4
setuptools>=69.0.2

# Social-Analyzer Dependencies
tld>=0.13
termcolor>=2.4.0
langdetect>=1.0.9
galeodes>=0.0.5

# Spiderfoot Dependencies
adblockparser>=0.7
ExifRead>=3.0.0
CherryPy>=18.8.0
cherrypy-cors>=1.7.0
Mako>=1.3.0
pysocks>=1.7.1
ipwhois>=1.2.0
ipaddr>=2.2.0
phonenumbers>=8.13.27
pygexf>=0.2.2
PyPDF2>=3.0.1
secure>=0.3.0
python-docx>=1.1.0
python-pptx>=0.6.23
publicsuffixlist>=0.10.0
openpyxl>=3.1.2

# Integration Dependencies for Tool Wrappers
dataclasses-json==0.6.4
typing-extensions==4.8.0

# Enhanced API and Webhook Support
fastapi-users>=12.1.3
fastapi-limiter>=0.1.5
fastapi-cache2>=0.2.1
starlette>=0.27.0

# Workflow and Automation (n8n integration)
n8n-python-sdk>=0.1.0
workflow-engine>=1.0.0
apscheduler>=3.10.4
celery-beat>=2.4.0

# Data Processing and Analysis
pandas>=2.1.4
numpy>=1.26.2
scipy>=1.11.4
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0

# Cloud and Containerization Support
docker>=6.1.3
kubernetes>=28.1.0
boto3>=1.34.0
azure-identity>=1.15.0
google-cloud-core>=2.4.1

# Enhanced Security Tools
nmap>=7.94
hashlib-compat>=1.0.1
cryptodome>=3.19.0
pycrypto>=2.6.1

# Development and Debugging Tools
ipython>=8.18.1
jupyter>=1.0.0
notebook>=7.0.6
debugging>=2.3.2
profiling>=0.1.7
memory-profiler>=0.61.0

# Additional Utilities
tqdm>=4.66.1
click-spinner>=0.1.10
spinners>=0.0.24
halo>=0.0.31
wasabi>=1.1.2
typer>=0.9.0

# Web Scraping and Parsing
beautifulsoup4>=4.11.2
lxml>=4.9.2
bs4

# Security and Cryptography
cryptography>=38.0.0
pyOpenSSL>=21.0.0
certifi

# Network and DNS
dnspython>=2.3.0
python-nmap>=0.7.1
netaddr>=0.8.0
aiohttp
whois
python-whois>=0.7.3

# ARGUS Dependencies
aioquic
gevent
pyfiglet
scapy

# Kraken Dependencies
paramiko>=2.12.0
pywifi
pyrad
ldap3

# RedTeam-ToolKit Dependencies
django>=4.1.9
gunicorn==20.1.0
bandit==1.7.4
bcrypt>=4.0.1
chardet==5.1.0
charset-normalizer>=2.0.12
contourpy>=1.0.6
cycler>=0.11.0
defusedxml>=0.7.1
dirsearch>=0.4.3
fonttools>=4.38.0
fpdf>=1.7.2
gitdb>=4.0.10
GitPython>=3.1.30
google>=3.0.0
greenlet>=2.0.1
humanize>=4.4.0
idna>=3.4
MarkupSafe>=2.1.1
matplotlib>=3.6.2
mypy-extensions>=0.4.3
networkx>=2.8.8
ntlm-auth>=1.5.0
numpy>=1.24.0
packaging>=22.0
pbr>=5.11.0
Pillow>=9.3.0
proxy-db>=0.3.1
pycparser>=2.21
PyNaCl>=1.5.0
pyparsing>=3.0.9
PySocks>=1.7.1
python-dateutil>=2.8.2
python-decouple>=3.6
pytz>=2022.7
PyYAML>=6.0
requests-mock>=1.10.0
requests-ntlm>=1.1.0
six>=1.16.0
smmap>=5.0.0
soupsieve>=2.3.2
SQLAlchemy>=2.0.1
sqlparse>=0.4.4
stevedore>=4.1.1
toml>=0.10.2
tomli>=2.0.1
typing_extensions>=4.4.0
uro>=0.0.4
setuptools>=65.5.1

# Social-Analyzer Dependencies
tld
termcolor
langdetect
galeodes

# Spiderfoot Dependencies
adblockparser>=0.7
ExifRead>=2.3.2
CherryPy>=18.8.0
cherrypy-cors>=1.6
Mako>=1.2.4
pysocks>=1.7.1
ipwhois>=1.1.0
ipaddr>=2.2.0
phonenumbers>=8.13.6
pygexf>=0.2.2
PyPDF2>=1.28.6
secure>=0.3.0
python-docx>=0.8.11
python-pptx>=0.6.21
publicsuffixlist>=0.9.3
openpyxl>=3.1.1

# Integration Dependencies for Tool Wrappers
dataclasses-json==0.6.4
typing-extensions==4.9.0

# Additional Security Tools
nmap
